requirements.txt
amqp
billiard
celery
certifi
chardet
click
credentials
Flask
idna
importlib-metadata
itsdangerous
Jinja2
joblib
kombu
MarkupSafe
numpy
opencv-python
phue
python-dotenv
pytz
requests
scikit-learn
scipy
six
spotipy
threadpoolctl
urllib3
vine
Werkzeug
zipp
